requirements.txt
firebase_admin==6.0.1
Flask==2.2.2
Flask_RESTful==0.3.9
python-dotenv==0.21.0
